[dart] type improvements (#8111)
* [dart] Configure and use import mapping

This prevents models from being generated which would clash with exisiting dart types, e.g. List.

* [dart] Fix decimal format not supported

* [dart-dio] Remove redundant modelToIgnore & ignore dart:core import

* modelToIgnore is now handled via importMappings the same way other generators do this
* choose not to import dart:core as this is available by default

* [dart-dio] EnumClass is a reserved word in built_value

* Review changes

* Fix regenerate docs

+++ samples/openapi3/client/petstore/dart-dio/petstore_client_lib_fake/lib/model/format_test.dart
@@ -1,5 +1,4 @@
 import 'dart:typed_data';
-import 'package:openapi/model/decimal.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
 
@@ -34,7 +33,7 @@ abstract class FormatTest implements Built<FormatTest, FormatTestBuilder> {
     
     @nullable
     @BuiltValueField(wireName: r'decimal')
-    Decimal get decimal;
+    double get decimal;
     
     @nullable
     @BuiltValueField(wireName: r'string')
